接続時間がオーバーフローしている問題を修正

+++ src/main/scala/com/github/unchama/seichiassist/database/manipulators/PlayerDataManipulator.scala
@@ -243,7 +243,7 @@ class PlayerDataManipulator(private val gateway: DatabaseGateway) {
       gateway.executeQuery(command).recordIteration { lrs =>
         val rankdata = new RankData()
         rankdata.name = lrs.getString("name")
-        rankdata.playtick = lrs.getInt("playtick")
+        rankdata.playtick = lrs.getLong("playtick")
         ranklist += rankdata
       }
     } catch {
